fix(api_server): error xml format fix
```
λ  playground git:(main) ✗ aws s3 mb s3://a
make_bucket failed: s3://a An error occurred (InvalidBucketName) when calling the CreateBucket operation: The specified bucket is not valid.
λ  playground git:(main) ✗ aws s3api create-bucket --bucket aaa_bbb
An error occurred (InvalidBucketName) when calling the CreateBucket operation: The specified bucket is not valid.

λ  playground git:(main) ✗ aws s3 mb s3://my-bucket
make_bucket: my-bucket
λ  playground git:(main) ✗ aws s3 mb s3://my-bucket
make_bucket failed: s3://my-bucket An error occurred (BucketAlreadyOwnedByYou) when calling the CreateBucket operation: The bucket that you tried to create already exists, and you own it.
```