fix(deploy): fix wrong num_nvme_disks for bss_server
Also did a clean up so that num_nvme_disks does not needed to be passed
into fractalbits-bootstrap any more. The origin idea of using it as
parameter is to verify the system before bootstrapping, however we can
always do the check from cloud-init logs for the same purpose to make dev
work simpler.